requirements_fastapi.txt
# Mini-RAG System - Python Dependencies
# Compatible with Python 3.9+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# PDF Processing
pdfplumber==0.11.4
PyPDF2==3.0.1

# LangChain - Core RAG Components
langchain==0.1.20
langchain-community==0.0.38

# Vector Database
chromadb==0.4.24

# Google Generative AI
google-generativeai==0.3.2
google-api-core==2.15.0

# Utilities
numpy==1.26.2
typing-extensions==4.9.0

# Optional: Better error messages
rich==13.7.0
